remove: 불필요한 import 및 타입 삭제
